style: rename src/features/home/elements -> src/features/home/components
